added @QueryField for reflection use
I created the @QueryField annotation to signify that a field in an
ApiRequest should be added to the query (in the final url). Its encode
method (defaults to true) can be set to false so that
ApiRequest#reflectQuery, which finds the query fields through reflection
and adds them to a QueryEncoder (a Map<String, String>), skips them.
Then they must manually be added to the query using protected
ApiRequest#modifyQuery(Map<String, String> query).

Reflection done using Type and the wrapper class Reflect with static
methods for Type.

Some utility methods in AnnotationUtils and RuntimeAnnotations (the
latter hasn't been used yet).