created RequestCache
RequestCache allows you to cache ApiRequests, putting them in internal
ConcurrentHashMaps. These maps map from url2id and id2path and vice
versa. The url is the unchanged url from the ApiRequest. The id is the
base64 url (file name safe) encoded SHA-1 (128 bits) hash of the url, a
unique, shorter (32 chars) id. The file name is the id, and the path is
constructed from the parent directory CACHE_DIRECTORY, an ApiRequest
subclass specific ApiRequest#getRelativePath, and the file name (id).
These maps are read/loaded from url2id.csv and id2path.csv in the
constructor and are saved there as well (overwriting) using
RequestCache#save. After the program runs and new ApiRequests are cached
using RequestCache#put(ApiRequest), the JVM shuts down. But I created a
SAVE_ON_EXIT Thread that calls requestCache.save() and added it to
Runtime.getRuntime()#addShutdownHook. Now whenever the JVM shuts down,
either normally as the main finishes or through System#exit, it saves
the requestCache first. However, if there is a JVM error or an OS
shutdown, it's not saved, but that shouldn't happen.