realized hotels.com actually has results in raw HTML, so it's only 3 sec
Changed HotelsHotelsScrapeRequest to extends HtmlRequest, not
RenderedHtmlRequest, because almost all the hotel information is
available in the raw html without javascript rendering.  I'm not sure,
but I think some of the info might be missing, like a rating here or
there, but it mostly works, and it only takes ~3 sec, so it should be
fast enough.